requirements.txt
discord.py
flask
requests
python-dotenv==1.0.1